feat: Divide company adjust form
