+++ src/components/organisms/index.ts
@@ -2,3 +2,4 @@ export { default as CompanyProductTable } from './CompanyProductTable';
 export { default as CompanyOrderTable } from './CompanyOrderTable';
 export { default as AdminProductTable } from './AdminProductTable';
 export { default as CompanyInfoForm } from './CompanyInfoForm';
+export { default as CompanyAdjustForm } from './CompanyAdjustForm';
